refactor(llm): split provider options from cache control
